+++ src/components/kytos/topology/Topology.vue
@@ -284,7 +284,7 @@ export default {
       })
     },
     project (lat, lng) {
-      return this.map.project(new mapboxgl.LngLat(+lng, +lat))
+      return this.map.project(new this.$mapboxgl.LngLat(+lng, +lat))
     },
     get_interface_owner (d) {
       /* Get the switch in which the "d" interface is connected */